LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/at-position-try-cssom-expected.txt


FAIL CSSPositionTryRule attribute values Can't find variable: CSSPositionTryRule
FAIL CSSPositionTryRule.style.setProperty setting allowed and disallowed properties assert_equals: expected 0 but got 999999
FAIL CSSPositionTryDescriptors.item undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.cssText undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-top) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-top) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-top] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-top] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginTop] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginTop] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-right) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-right) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-right] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-right] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginRight] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginRight] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-bottom) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-bottom) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-bottom] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-bottom] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginBottom] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginBottom] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-left) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-left) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-left] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-left] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginLeft] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginLeft] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-block) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-block) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-block] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-block] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginBlock] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginBlock] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-block-start) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-block-start) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-block-start] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-block-start] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginBlockStart] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginBlockStart] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-block-end) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-block-end) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-block-end] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-block-end] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginBlockEnd] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginBlockEnd] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-inline) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-inline) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-inline] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-inline] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginInline] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginInline] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-inline-start) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-inline-start) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-inline-start] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-inline-start] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginInlineStart] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginInlineStart] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(margin-inline-end) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(margin-inline-end) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-inline-end] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[margin-inline-end] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginInlineEnd] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[marginInlineEnd] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inset) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inset) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(top) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(top) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[top] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[top] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(left) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(left) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[left] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[left] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(right) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(right) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[right] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[right] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(bottom) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(bottom) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[bottom] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[bottom] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inset-block) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inset-block) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-block] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-block] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetBlock] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetBlock] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inset-block-start) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inset-block-start) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-block-start] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-block-start] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetBlockStart] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetBlockStart] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inset-block-end) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inset-block-end) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-block-end] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-block-end] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetBlockEnd] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetBlockEnd] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inset-inline) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inset-inline) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-inline] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-inline] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetInline] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetInline] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inset-inline-start) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inset-inline-start) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-inline-start] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-inline-start] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetInlineStart] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetInlineStart] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inset-inline-end) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inset-inline-end) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-inline-end] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-inline-end] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetInlineEnd] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetInlineEnd] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(width) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(width) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[width] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[width] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(height) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(height) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[height] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[height] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(min-width) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(min-width) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[min-width] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[min-width] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[minWidth] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[minWidth] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(max-width) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(max-width) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[max-width] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[max-width] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[maxWidth] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[maxWidth] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(min-height) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(min-height) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[min-height] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[min-height] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[minHeight] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[minHeight] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(max-height) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(max-height) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[max-height] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[max-height] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[maxHeight] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[maxHeight] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(block-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(block-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[block-size] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[block-size] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[blockSize] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[blockSize] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(min-block-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(min-block-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[min-block-size] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[min-block-size] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[minBlockSize] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[minBlockSize] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(max-block-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(max-block-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[max-block-size] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[max-block-size] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[maxBlockSize] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[maxBlockSize] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inline-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inline-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inline-size] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inline-size] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inlineSize] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inlineSize] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(min-inline-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(min-inline-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[min-inline-size] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[min-inline-size] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[minInlineSize] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[minInlineSize] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(max-inline-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(max-inline-size) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[max-inline-size] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[max-inline-size] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[maxInlineSize] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[maxInlineSize] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(place-self) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(place-self) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[place-self] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[place-self] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[placeSelf] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[placeSelf] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(align-self) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(align-self) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[align-self] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[align-self] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[alignSelf] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[alignSelf] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(justify-self) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(justify-self) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[justify-self] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[justify-self] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[justifySelf] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[justifySelf] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(position-anchor) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(position-anchor) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[position-anchor] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[position-anchor] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[positionAnchor] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[positionAnchor] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(inset-area) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(inset-area) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-area] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[inset-area] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetArea] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[insetArea] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(color) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(color) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[color] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[color] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(align-items) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(align-items) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[align-items] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[align-items] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[alignItems] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[alignItems] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(align-content) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(align-content) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[align-content] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[align-content] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[alignContent] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[alignContent] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(background) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(background) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[background] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[background] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(display) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(display) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[display] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[display] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(position) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(position) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[position] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[position] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(writing-mode) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(writing-mode) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[writing-mode] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[writing-mode] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[writingMode] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[writingMode] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(direction) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(direction) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[direction] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[direction] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.getPropertyValue(syntax) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors.setProperty(syntax) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[syntax] (set) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
FAIL CSSPositionTryDescriptors[syntax] (get) undefined is not an object (evaluating 'style.sheet.cssRules[0].style')
